Add cache management infrastructure to binary_distribution
This change adds a class which provides two levels of caching to the
module.

At one level, it caches buildcache index files from mirrors,
and uses a hash of the index contents to track whether the local caches
need to be updates.

Then the cached index files can be used to answer questions like "what
binary specs are available to me, based on my configured mirrors".  In
that case, the cached index files are used to populate the other level
of cache, the built spec cache.

The built spec cache in the cache manager replaces the previous global
"_cached_specs" variable with a cache that keeps track of which mirrors
contains which concrete specs.  However, if this cache does not contain
a target spec, the direct approach to fetching a spec.yaml from a mirror
is still supported.

One goal of this work is to support a new install option which will tell
spack to avoid installation from binary unless the full_hash of the binary
spec on te mirror mathes that of the local concrete spec.